backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0

sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0

chromadb>=0.5.0
rank-bm25>=0.2.0

pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.0
beautifulsoup4>=4.12.0

python-dotenv>=1.0.0
aiofiles>=23.0.0
